refactor(hickey): extract triggerDownload, dedupe anchor-click download
The createElement(a) -> href/download/rel -> append -> click -> remove
sequence was duplicated in sessionTransfer.ts and exportSessionAsHtml.ts.
Extract to download.ts; callers keep their own revoke timing.
